+++ NEWS.md
@@ -69,6 +69,9 @@
 * Setting `clean = TRUE` in `deploy_site_github()` removes old files from the 
   deployed site before building a new one (#1297).
 
+* Missing topics makes the build fail when the environment variable `CI` is set
+  to `"true"` (@ThierryO, #1378).
+
 # pkgdown 1.5.1
 
 * Syntax highlighting works on Windows once more (#1282).